+++ config/src/defaults.rs
@@ -144,6 +144,10 @@ performance = "High"
 # Note: You can set different font families but Rio terminal
 # will always look for regular font bounds whene
 #
+# You can also set family on root to overwritte all fonts
+# [fonts]
+#   family = "cascadiamono"
+#
 # Example
 #   [fonts]
 #   size = 16